+++ sdk/azcore/arm/runtime/pipeline.go
@@ -25,23 +25,12 @@ func NewPipeline(module, version string, cred azcore.TokenCredential, options *a
 	if len(ep) == 0 {
 		ep = arm.AzurePublicCloud
 	}
-	policies := []policy.Policy{}
-	if !options.Telemetry.Disabled {
-		policies = append(policies, azruntime.NewTelemetryPolicy(module, version, &options.Telemetry))
-	}
+	perCallPolicies := []policy.Policy{}
 	if !options.DisableRPRegistration {
-		regRPOpts := RegistrationOptions{
-			HTTPClient: options.Transport,
-			Logging:    options.Logging,
-			Retry:      options.Retry,
-			Telemetry:  options.Telemetry,
-		}
-		policies = append(policies, NewRPRegistrationPolicy(string(ep), cred, &regRPOpts))
+		regRPOpts := RegistrationOptions{ClientOptions: options.ClientOptions}
+		perCallPolicies = append(perCallPolicies, NewRPRegistrationPolicy(string(ep), cred, &regRPOpts))
 	}
-	policies = append(policies, options.PerCallPolicies...)
-	policies = append(policies, azruntime.NewRetryPolicy(&options.Retry))
-	policies = append(policies, options.PerRetryPolicies...)
-	policies = append(policies,
+	perRetryPolicies := []policy.Policy{
 		azruntime.NewBearerTokenPolicy(cred, azruntime.AuthenticationOptions{
 			TokenRequest: policy.TokenRequestOptions{
 				Scopes: []string{shared.EndpointToScope(string(ep))},
@@ -49,6 +38,6 @@ func NewPipeline(module, version string, cred azcore.TokenCredential, options *a
 			AuxiliaryTenants: options.AuxiliaryTenants,
 		},
 		),
-		azruntime.NewLogPolicy(&options.Logging))
-	return azruntime.NewPipeline(options.Transport, policies...)
+	}
+	return azruntime.NewPipeline(module, version, perCallPolicies, perRetryPolicies, &options.ClientOptions)
 }