+++ sdk/messaging/azservicebus/example_session_receiver_test.go
@@ -6,26 +6,7 @@ package azservicebus_test
 import (
 	"context"
 	"fmt"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
-	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
 )
-
-func init() {
-	initExamples()
-
-	if client != nil {
-		sender, err := client.NewSender(exampleSessionQueue)
-		exitOnError("Failed to create sender for session examples", err)
-
-		for i := 0; i < 2; i++ {
-			err = sender.SendMessage(context.Background(), &azservicebus.Message{
-				SessionID: to.StringPtr("Example Session ID"),
-			})
-			exitOnError("Failed to send example message to session", err)
-		}
-	}
-}
 
 func ExampleClient_AcceptSessionForQueue() {
 	sessionReceiver, err := client.AcceptSessionForQueue(context.TODO(), "exampleSessionQueue", "Example Session ID", nil)